chore(frontEnd): no-delete' rule in Prettier configuration
This commit disables the 'no-delete' rule in the Prettier configuration
file (.prettierrc) to allow the use of the 'delete' operator in JavaScript
and TypeScript code. This rule was causing conflicts with certain code
patterns and removing flexibility in code formatting.

Changes:
- Disabled the 'no-delete' rule in the Prettier configuration file
- Allows the use of the 'delete' operator in code

This change ensures a smoother development experience and greater
flexibility in code formatting.