refactor(frontEnd): Replaced the <img> tag with the Next.js <Image> component
In this commit, we replaced the use of the conventional <img> tag with the <Image> component provided by Next.js. The <Image> component provides automatic optimizations for image loading, including delayed loading and image loading optimized for small screen devices.

This replacement aims to improve performance and user experience in our application by ensuring images load efficiently and responsively.

It is important to note that the <Image> component requires the use of the 'next/image' module and has some differences in relation to the <img> tag, such as the need to explicitly specify width and height. However, the performance and usability benefits outweigh these small usage differences.

This update should contribute to a faster and more responsive user experience in our application.